requirements.txt
tensorflow>=2.13.0
gradio>=4.0.0
pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
